fix: `forall` and `map` fixed
forall) didnt check element is in given set. `else if (!p(a)) false` --> `else if (s(a) && !p(a)) false`. if existence has not checked, any element that does not satispy `p` will result false even though it is not contained in `s`.

map) element in set will be mapped with `f` so we need contravariant `f^-1` to use `compose` keyword(`s compose (f^-1)`). f may not have inverse. instead, just check existence.